Репозиторий для получения данных по сети для кредитов и вкладов

+++ app/src/main/java/com/example/sandboxbank/App/core/deposit/data/db/FinancialEntity.kt
@@ -7,12 +7,12 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "financial_items", indices = [Index(value = ["type"])])
 data class FinancialEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     val type: String,
     val openDate: Long,
     val percentRate: Double,
-    val percentType: Int,
-    val period: Int,
-    val balance: Double,
+    val percentType: Long,
+    val period: Long,
+    val balance: Long,
     val name: String
 )
